pyproject.toml
[project]
name = "meet-your-new-client"
version = "0.2.0"
description = "Experiment about markdown document conversion for RAG on reports"
readme = "README.md"
requires-python = "==3.12.8"
dependencies = [
    "markitdown[all]==0.1.2",
    "tqdm>=4.67.1",
    "polars>=1.22.0",
    "pymupdf>=1.25.3",
    "jsonschema>=4.23.0",
    "pyprojroot>=0.3.0",
    "statsmodels>=0.14.4",
    "hydra-core>=1.3.2",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "pyarrow>=19.0.1",
    "matplotlib>=3.10.1",
    "python-pptx>=1.0.2",
    "hydra-list-sweeper>=1.1.0",
    "nbclient>=0.10.2",
    "plotnine>=0.14.6",
    "great-tables>=0.17.0",
    "litellm[proxy]>=1.71.1",
    "modal>=0.75.6",
    "tenacity>=9.1.2",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "nbformat>=5.10.4",
    "pytest>=8.3.5",
    "ruff>=0.11.5",
]
experiment = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
